cart-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-settings==2.5.2
asyncpg==0.29.0
redis==5.0.8
aiokafka==0.10.0
python-dotenv==1.0.1
structlog==24.4.0
pytest==8.3.3
pytest-asyncio==0.24.0

